Additional Methods

Additional methods

All ReadDatastores permit construction from fastq files, iteration, and indexing.

But some specific ReadDatastore types also have additional specific methods. They are listed here.

ReadDatastores.read_tag — Function.

read_tag(lr::LinkedReads, i::Integer)

Get the tag for the i'th read

source

ReadDatastores.name — Function.

Get the name of the datastore

source

ReadDatastores.maxseqlen — Function.

Get the length of the longest sequence in the datastore

source

Get the length of the longest sequence in the datastore

source

ReadDatastores.orientation — Function.

Get the orientation of the read pairs

source

Missing docstring.

Missing docstring for stream. Check Documenter's build log for details.