Building & loading datastores

Building Read Datastores

To build a read datastore you first need to decide what sort of read data you have.

If you have short (somewhat regular in length) paired reads such as that were sequenced from a paired-end library or a long mate pair library, you should build a PairedReads datastore.

If you have long reads of inconsistent length, such as you would get from single-molecule real-time sequencing of a long read library, you should build a LongReads datastore.

If you have short (somewhat regular in length) paired reads, and thr pairs are linked linked through some proximity tagging mechanism (like 10x), then you should build a LinkedReads datastore.

The process of building these datastores is consistent, but for each datastore there are datastore specific options the constructor will accept:

ReadDatastores.PairedReads — Type.

PairedReads(rdrx::FASTQ.Reader, rdry::FASTQ.Reader, outfile::String, name::String, minsize::Integer, maxsize::Integer, fragsize::Integer, orientation::PairedReadOrientation)

Construct a Paired Read Datastore from a pair of FASTQ file readers.

Paired-end sequencing reads typically come in the form of two FASTQ files, often named according to a convention *_R1.fastq and *_R2.fastq. One file contains all the "left" sequences of each pair, and the other contains all the "right" sequences of each pair. The first read pair is made of the first record in each file.

Arguments

rdrx::FASTQ.Reader: The reader of the *_R1.fastq file.

rdxy::FASTQ.Reader: The reader of the *_R2.fastq file.

outfile::String: A prefix for the datastore's filename, the full filename will include a ".prseq" extension, which will be added automatically.

name::String: A string denoting a default name for your datastore. Naming datastores is useful for downstream applications.

minsize::Integer: A minimum read length (in base pairs). When building the datastore, if any pair of reads has one or both reads shorter than this cutoff, then the pair will be discarded.

maxsize::Integer: A maximum read length (in base pairs). When building the datastore, if any read has a greater length, it will be resized to this maximum length and added to the datastore.

fragsize::Integer: The average fragment length of the paired end library that was sequenced. This value is entirely optional, but may be important for downstream applications.

orientation::PairedReadOrientation: The orientation of the reads. Set it to FwRv for building a datastore from a sequenced paired end library, and set it to RvFw if you are building the datastore from reads sequenced from a long mate pair library.

Examples

julia> using FASTX, ReadDatastores

julia> fwq = open(FASTQ.Reader, "test/ecoli_tester_R1.fastq")
FASTX.FASTQ.Reader{TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}}(BioGenerics.Automa.State{TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}}(TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}(<mode=idle>), 1, 1, false), nothing)

julia> rvq = open(FASTQ.Reader, "test/ecoli_tester_R2.fastq")
FASTX.FASTQ.Reader{TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}}(BioGenerics.Automa.State{TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}}(TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}(<mode=idle>), 1, 1, false), nothing)

julia> ds = PairedReads(fwq, rvq, "ecoli-test-paired", "my-ecoli-test", 250, 300, 0, FwRv)
[ Info: Building paired read datastore from FASTQ files
[ Info: Writing paired reads to datastore
[ Info: Done writing paired read sequences to datastore
[ Info: 0 read pairs were discarded due to a too short sequence
[ Info: 14 reads were truncated to 300 base pairs
[ Info: Created paired sequence datastore with 10 sequence pairs
Paired Read Datastore 'my-ecoli-test': 20 reads (10 pairs)

source

ReadDatastores.LongReads — Type.

LongReads(rdr::FASTQ.Reader, outfile::String, name::String, min_size::Integer)

Construct a Long Read Datastore from a FASTQ file reader.

Arguments

rdr::FASTQ.Reader: The reader of the fastq formatted file.

outfile::String: A prefix for the datastore's filename, the full filename will include a ".loseq" extension, which will be added automatically.

name::String: A string denoting a default name for your datastore. Naming datastores is useful for downstream applications.

min_size::Integer: A minimum read length (in base pairs). When building the datastore, if any read sequence is shorter than this cutoff, then the read will be discarded.

Examples

julia> using FASTX, ReadDatastores

julia> longrdr = open(FASTQ.Reader, "test/human_nanopore_tester_2D.fastq")
FASTX.FASTQ.Reader{TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}}(BioGenerics.Automa.State{TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}}(TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}(<mode=idle>), 1, 1, false), nothing)

julia> ds = LongReads(longrdr, "human-nanopore-tester", "nanopore-test", 0)
[ Info: Building long read datastore from FASTQ file
[ Info: Writing long reads to datastore
[ Info: Done writing paired read sequences to datastore
[ Info: 0 reads were discarded due to a too short sequence
[ Info: Writing index to datastore
[ Info: Built long read datastore with 10 reads
Long Read Datastore 'nanopore-test': 10 reads

source

ReadDatastores.LinkedReads — Type.

LinkedReads(fwq::FASTQ.Reader, rvq::FASTQ.Reader, outfile::String, name::String, format::LinkedReadsFormat, readsize::Integer, chunksize::Int = 1000000)

Construct a Linked Read Datastore from a pair of FASTQ file readers.

Paired sequencing reads typically come in the form of two FASTQ files, often named according to a convention *_R1.fastq and *_R2.fastq. One file contains all the "left" sequences of each pair, and the other contains all the "right" sequences of each pair. The first read pair is made of the first record in each file.

Arguments

rdrx::FASTQ.Reader: The reader of the *_R1.fastq file.

rdxy::FASTQ.Reader: The reader of the *_R2.fastq file.

outfile::String: A prefix for the datastore's filename, the full filename will include a ".prseq" extension, which will be added automatically.

name::String: A string denoting a default name for your datastore. Naming datastores is useful for downstream applications.

format::LinkedReadsFormat: Specify the linked reads format of your fastq files. If you have plain 10x files, set format to Raw10x. If you have 10x reads output in the UCDavis format, set the format to UCDavis10x.

chunksize::Int = 1000000: How many read pairs to process per disk batch during the tag sorting step of construction.

Examples

julia> using FASTX, ReadDatastores

julia> fqa = open(FASTQ.Reader, "test/10x_tester_R1.fastq")
FASTX.FASTQ.Reader{TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}}(BioGenerics.Automa.State{TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}}(TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}(<mode=idle>), 1, 1, false), nothing)

julia> fqb = open(FASTQ.Reader, "test/10x_tester_R2.fastq")
FASTX.FASTQ.Reader{TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}}(BioGenerics.Automa.State{TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}}(TranscodingStreams.TranscodingStream{TranscodingStreams.Noop,IOStream}(<mode=idle>), 1, 1, false), nothing)

julia> ds = LinkedReads(fqa, fqb, "10xtest", "ucdavis-test", UCDavis10x, 250)
[ Info: Building tag sorted chunks of 1000000 pairs
[ Info: Dumping 83 tag-sorted read pairs to chunk 0
[ Info: Dumped
[ Info: Processed 83 read pairs so far
[ Info: Finished building tag sorted chunks
[ Info: Performing merge from disk
[ Info: Leaving space for 83 read_tag entries
[ Info: Chunk 1 is finished
[ Info: Finished merge from disk
[ Info: Writing down 83 read tag entries
[ Info: Created linked sequence datastore with 83 sequence pairs
Linked Read Datastore 'ucdavis-test': 166 reads (83 pairs)

source

Loading Read Datastores

Once you have built a datastore, you can open it in other projects again using a Base.open method:

julia> ds = open(PairedReads, "test/ecoli-pe.prseq", "my-ecoli-pe")
Paired Read Datastore 'my-ecoli-pe': 20 reads (10 pairs)

The open method takes a ReadDatastore type, the filename of the datastore, and, optionally a name for the datastore, which may be omitted, in which case the datastore will use the default name that was specified on construction.